pyproject.toml
[project]
name = "kpub"
version = "0.1.0"
description = "Automatic publication classifier for W. M. Keck Observatory"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=3.0.1",
    "pymongo>=4.16.0",
    "PyMuPDF>=1.25.0",
    "python-dotenv>=1.2.1",
    "PyYAML>=6.0",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
]
